fp8/science/fp8/tensor.py
import torch
from torch.utils._pytree import tree_map

class FP8Meta:
    # Placeholder class for FP8 metadata
    pass

def _fp8_to_func():
    pass

class DispatchWrapperTensor(torch.Tensor):
    SUPPORTED_OPS = {}

    @classmethod
    def add_impl(cls, func, impl):
        cls.SUPPORTED_OPS[func] = impl

    __torch_function__ = torch._C._disabled_torch_function_impl

    @classmethod
    def __torch_dispatch__(cls, func, types, args=(), kwargs=None):
        func_name = f"{func.__module__}.{func.__name__}"
        if func in cls.SUPPORTED_OPS:
            res = cls.SUPPORTED_OPS[func](*args, **kwargs)
            if res is not NotImplemented:
                return res

        if not hasattr(cls, "wrap") or not hasattr(cls, "unwrap"):
            return NotImplemented

        res = func(*tree_map(cls.unwrap, args), **tree_map(cls.unwrap, kwargs or {}))
        return tree_map(cls.wrap, res)

class FP8Tensor(DispatchWrapperTensor):
    @classmethod
    def get_wrapper_properties(cls, tensor):
        return tensor, {"size": tensor.size(), "dtype": tensor.dtype, "device": tensor.device}

    def __init__(self, tensor, **kwargs):
        self._tensor = self.quantize_to_fp8(tensor)

    @staticmethod
    def quantize_to_fp8(tensor):
        return torch.randint_like(tensor, low=0, high=10, dtype=torch.int8)

    @staticmethod
    def dequantize_from_fp8(int_data):
        return int_data.to(torch.float32)

    @classmethod
    def unwrap(cls, e):
        if isinstance(e, cls):
            return cls.dequantize_from_fp8(e._tensor)
        else:
            return e

    @classmethod
    def wrap(cls, e):
        if isinstance(e, torch.Tensor) and e.dtype == torch.float32:
            return cls(e)
        else:
            return e

    @classmethod
    def __torch_dispatch__(cls, func, types, args=(), kwargs=None):
        func_name = func.__name__

        if func_name == '_to_copy':
            # Handle 'to' operation
            self = args[0]
            dtype = kwargs.get('dtype', None)
            if dtype == torch.float32 or dtype is None:
                if isinstance(self, FP8Tensor):
                    # Dequantize
                    dequantized = cls.dequantize_from_fp8(self._int_data, self.scale)
                    return dequantized
                else:
                    return func(*args, **kwargs)
            else:
                raise NotImplementedError("FP8Tensor only supports conversion to float32")
        elif func_name == 'clone':
            # Handle clone operation
            self = args[0]
            if isinstance(self, FP8Tensor):
                cloned = FP8Tensor(self._tensor.clone())
                return cloned
            else:
                return func(*args, **kwargs)
        elif func_name == 'zero_':
            # Handle zero_ operation
            self = args[0]
            if isinstance(self, FP8Tensor):
                self._tensor.zero_()
                return self
            else:
                return func(*args, **kwargs)
        else:
            # For other operations, dequantize, perform op, then re-quantize
            res = func(*tree_map(cls.unwrap, args), **tree_map(cls.unwrap, kwargs or {}))
            return tree_map(cls.wrap, res)

    @property
    def data(self):
        return self._tensor

    @data.setter
    def data(self, new_data):
        if isinstance(new_data, FP8Tensor):
            self._tensor = new_data.data
        elif isinstance(new_data, torch.Tensor):
            self._tensor = self.quantize_to_fp8(new_data)
        else:
            raise TypeError("Expected new_data to be FP8Tensor or torch.Tensor")

    def zero_(self):
        self._tensor.zero_()
        return self

    # def clone(self):
    #     cloned = FP8Tensor(self.orig_data.clone())
    #     return cloned

    # @classmethod
    # def from_metadata(cls, tensor, fp8_meta):
    #     obj = cls.__new__(cls)
    #     obj.orig_data = tensor
    #     obj.fp8_meta = fp8_meta
    #     obj._int_data, obj.scale = cls.quantize_to_fp8(tensor)
    #     obj.dtype = tensor.dtype
    #     return obj

    def __repr__(self):
        return f"FP8Tensor({self._tensor})"

# Note: You would need to adjust the import statements and any other dependencies according to your environment.

# Implementing test functions would involve writing unit tests that utilize the FP8Tensor class.


if __name__ == "__main__":
    fp8_tensor = FP8Tensor(torch.randn((4, 4)))
